refactor(platform): move service management impl
